Move and rename students/-schema to src/schemas/search-schemas.ts
